requirements.txt
Django==4.2
djangorestframework==3.14
djangorestframework-simplejwt==5.2
drf-spectacular==0.26.5
python-dotenv==1.0
setuptools==68.0.0
